+++ src/ops/subdivide.rs
@@ -189,7 +189,7 @@ impl MeshGraph {
             normals.insert(center_v, (start_normal + end_normal).normalize());
         }
 
-        let new_he = self.add_halfedge(center_v, end_v);
+        let new_he = self.add_halfedge(center_v, end_v)?;
         // inserted just above
         self.vertices[center_v].outgoing_halfedge = Some(new_he);
 
@@ -199,7 +199,7 @@ impl MeshGraph {
             added_halfedges.push(new_face_he);
         }
 
-        let new_twin = self.add_halfedge(center_v, start_v);
+        let new_twin = self.add_halfedge(center_v, start_v)?;
 
         if let Some(new_face_he) = self.subdivide_face(twin_id, new_twin, center_v) {
             added_halfedges.push(new_face_he);
@@ -257,13 +257,13 @@ impl MeshGraph {
             .or_else(error_none!("Last halfedge not found"))?;
 
         // rewire existing face
-        let new_he = self.add_halfedge(center_v, self.halfedges[next_he].end_vertex);
+        let new_he = self.add_halfedge(center_v, self.halfedges[next_he].end_vertex)?;
 
         self.halfedges[existing_halfedge_id].next = Some(new_he);
         self.halfedges[new_he].next = Some(last_he);
         self.halfedges[new_he].face = Some(face_id);
 
-        let new_twin = self.add_halfedge(self.halfedges[next_he].end_vertex, center_v);
+        let new_twin = self.add_halfedge(self.halfedges[next_he].end_vertex, center_v)?;
 
         // insert new face
         let new_face_id = self.add_face(new_halfedge_id, next_he, new_twin);