+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2026-06-17
+
+- Updated dependencies rerun, parry3d and glam
+- Fixed `merge_one_ring` edge cases
+- A bunch of `edge` methods now return `Option<...>` because they can fail if given invalid `VertexId`s.
+- Added `MeshGraph::merge_vertices`
+
 ## [0.6.0] - 2026-04-10
 
 - Updated dependencies rerun and hashbrown